Exp 6 Chatbot/backend/requirements.txt
fastapi
uvicorn[standard]
langgraph
openai
python-dotenv
httpx
pydantic
